+++ eth-types/Cargo.toml
@@ -18,7 +18,6 @@ serde_with = "1.12"
 uint = "0.9.1"
 itertools = "0.10"
 libsecp256k1 = "0.7"
-#keccak256 = { path = "../keccak256"}
 subtle = "2.4"
 sha3 = "0.10"
 num = "0.4"
