+++ zkevm-circuits/src/sig_circuit/utils.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use eth_types::{sign_types::SignData, Field};
+use eth_types::Field;
 use halo2_base::{AssignedValue, QuantumCell};
 use halo2_ecc::{
     bigint::CRTInteger,
@@ -11,8 +11,6 @@ use halo2_proofs::{
     circuit::{Cell, Value},
     halo2curves::secp256k1::{Fp, Fq},
 };
-
-use crate::witness::Transaction;
 
 // Hard coded parameters.
 // FIXME: allow for a configurable param.
@@ -146,36 +144,4 @@ pub(crate) fn pub_key_hash_to_address<F: Field>(pk_hash: &[u8]) -> F {
     pk_hash[32 - 20..]
         .iter()
         .fold(F::zero(), |acc, b| acc * F::from(256) + F::from(*b as u64))
-}
-
-pub(crate) fn get_sign_data(
-    txs: &[Transaction],
-    max_txs: usize,
-    chain_id: usize,
-) -> Result<Vec<SignData>, halo2_proofs::plonk::Error> {
-    let padding_txs = (txs.len()..max_txs)
-        .into_iter()
-        .map(|i| {
-            let mut tx = Transaction::dummy(chain_id as u64);
-            tx.id = i + 1;
-            tx
-        })
-        .collect::<Vec<Transaction>>();
-    let signatures: Vec<SignData> = txs
-        .iter()
-        .chain(padding_txs.iter())
-        .map(|tx| {
-            if tx.tx_type.is_l1_msg() {
-                // dummy signature
-                Ok(SignData::default())
-            } else {
-                // TODO: map err or still use bus_mapping::err?
-                tx.sign_data().map_err(|e| {
-                    log::error!("tx_to_sign_data error for tx {:?}", e);
-                    halo2_proofs::plonk::Error::Synthesis
-                })
-            }
-        })
-        .collect::<Result<Vec<SignData>, halo2_proofs::plonk::Error>>()?;
-    Ok(signatures)
 }
