+++ zkevm-circuits/src/tx_circuit/test.rs
@@ -8,8 +8,7 @@ use std::cmp::max;
 
 use super::*;
 use crate::{
-    sig_circuit::get_sign_data,
-    tx_circuit::dev::TxCircuitTester,
+    tx_circuit::{dev::TxCircuitTester, get_sign_data},
     util::{log2_ceil, unusable_rows},
 };
 use eth_types::{address, evm_types::gas_utils::tx_data_gas_cost, word, H256, U256, U64};
